renamed interface
