+++ scrapper/src/main/java/edu/java/api/controllers/ChatController.java
@@ -1,6 +1,6 @@
 package edu.java.api.controllers;
 
-import edu.java.api.services.ChatService;
+import edu.java.api.services.ChatsService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping(path = "/tg-chat")
 @RequiredArgsConstructor
 public class ChatController {
-    private final ChatService service;
+    private final ChatsService service;
 
     @PostMapping(path = "/{id}")
     @Operation(summary = "Зарегистрировать чат")
