+++ scrapper/src/main/java/edu/java/api/services/ChatsService.java
@@ -1,6 +1,6 @@
 package edu.java.api.services;
 
-public interface ChatService {
+public interface ChatsService {
     void registerChat(long chatId);
 
     void deleteChat(long chatId);
